+++ src/commands/project/deploy/resume.ts
@@ -12,7 +12,13 @@ import { Duration } from '@salesforce/kit';
 import { DeployStages } from '../../../utils/deployStages.js';
 import { DeployResultFormatter } from '../../../formatters/deployResultFormatter.js';
 import { API, DeployResultJson } from '../../../utils/types.js';
-import { buildComponentSet, determineExitCode, executeDeploy, isNotResumable } from '../../../utils/deploy.js';
+import {
+  buildComponentSet,
+  buildDeployUrl,
+  determineExitCode,
+  executeDeploy,
+  isNotResumable,
+} from '../../../utils/deploy.js';
 import { DeployCache } from '../../../utils/deployCache.js';
 import { DEPLOY_STATUS_CODES_DESCRIPTIONS } from '../../../utils/errorCodes.js';
 import { coverageFormattersFlag } from '../../../utils/flags.js';
@@ -78,6 +84,8 @@ export default class DeployMetadataResume extends SfCommand<DeployResultJson> {
 
   public static errorCodes = toHelpSection('ERROR CODES', DEPLOY_STATUS_CODES_DESCRIPTIONS);
 
+  private deployUrl?: string;
+
   public async run(): Promise<DeployResultJson> {
     const [{ flags }, cache] = await Promise.all([this.parse(DeployMetadataResume), DeployCache.create()]);
     const jobId = cache.resolveLatest(flags['use-most-recent'], flags['job-id'], true);
@@ -86,11 +94,11 @@ export default class DeployMetadataResume extends SfCommand<DeployResultJson> {
     const deployOpts = { ...cache.maybeGet(jobId), async: false };
 
     let result: DeployResult;
+    const org = await Org.create({ aliasOrUsername: deployOpts['target-org'] });
 
     // If we already have a status from cache that is not resumable, display a warning and the deploy result.
     if (isNotResumable(deployOpts.status)) {
       this.warn(messages.getMessage('warning.DeployNotResumable', [jobId, deployOpts.status]));
-      const org = await Org.create({ aliasOrUsername: deployOpts['target-org'] });
       const componentSet = await buildComponentSet({ ...deployOpts, wait: Duration.seconds(0) });
       const mdapiDeploy = new MetadataApiDeploy({
         // setting an API version here won't matter since we're just checking deploy status
@@ -122,11 +130,19 @@ export default class DeployMetadataResume extends SfCommand<DeployResultJson> {
         this.project,
         jobId
       );
-
+      this.deployUrl = buildDeployUrl(org, jobId);
       new DeployStages({
         title: 'Resuming Deploy',
         jsonEnabled: this.jsonEnabled(),
-      }).start({ deploy, username: deployOpts['target-org'] });
+      }).start({
+        deploy,
+        username: deployOpts['target-org'],
+        ...(flags.verbose
+          ? {
+              deployUrl: this.deployUrl,
+            }
+          : {}),
+      });
 
       result = await deploy.pollStatus(500, wait.seconds);
 
@@ -148,6 +164,13 @@ export default class DeployMetadataResume extends SfCommand<DeployResultJson> {
 
     if (!this.jsonEnabled()) formatter.display();
 
-    return formatter.getJson();
+    return this.mixinUrlMeta(await formatter.getJson());
+  }
+
+  private mixinUrlMeta(json: DeployResultJson): DeployResultJson {
+    if (this.deployUrl) {
+      json.deployUrl = this.deployUrl;
+    }
+    return json;
   }
 }